refactor: restructure filenames and update imports
