+++ src/components/auth/register-form.tsx
@@ -1,27 +1,18 @@
 "use client";
 
-import {useState} from "react";
+import React, {useState} from "react";
 import {useRouter} from "next/navigation";
 import Link from "next/link";
 import {cn} from "@/lib/utils";
 import {Button} from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import {Card, CardContent, CardDescription, CardHeader, CardTitle} from "@/components/ui/card";
 import {Input} from "@/components/ui/input";
 import {Label} from "@/components/ui/label";
-import {useAuthContext} from "@/contexts/AuthContext";
+import {useAuthContext} from "@/contexts/auth-context";
 import {Alert, AlertDescription} from "@/components/ui/alert";
 import {toast} from "sonner";
 
-export function RegisterForm({
-                               className,
-                               ...props
-                             }: React.ComponentPropsWithoutRef<"div">) {
+export function RegisterForm({className, ...props}: React.ComponentPropsWithoutRef<"div">) {
   const router = useRouter();
   const {register, error, isLoading} = useAuthContext();
   const [email, setEmail] = useState("");
@@ -49,12 +40,7 @@ export function RegisterForm({
       return;
     }
 
-    const success = await register({
-      email,
-      username,
-      password,
-      password_confirm: passwordConfirm,
-    });
+    const success = await register({email, username, password, password_confirm: passwordConfirm});
 
     if (success) {
       toast.success("Register successful");
