+++ src/components/sidebar/nav-user.tsx
@@ -26,7 +26,7 @@ import {
   useSidebar,
 } from "@/components/ui/sidebar";
 
-import { useAuthContext } from "@/contexts/AuthContext";
+import { useAuthContext } from "@/contexts/auth-context";
 import { useRouter } from "next/navigation";
 
 export function NavUser() {
